Smoothing off rough edges in preparation for Copper 0.7 release:
Updated Javadoc and other documentation to match the new interface and functionality.

Added an automatic versioning apparatus to the ANT build that makes the version number, the build's Mercurial changeset ID, and the build time accessible to Copper itself. This version information now appears in comments at the top of all parser source files generated by the non-legacy pipeline, and is also accessible through a new switch '-version'.

Changed the '-compose' switch to '-mda' and added a boolean parameter for it to the ANT task, with corresponding changes to ParserCompilerParameters and other classes.

--HG--
branch : 0.7-devel